Improved parsing and added support for ls and cd
Parsing is a bit more general. Also added support for ls and cd but it
makes bad assumptions. Such as the args to cd will be nice and it does
the call for ls manually. This all needs to be made general and there
should be no assumptions for the input.