Started working with the interupt code
Have code for how to handle timer interupts relative to how we will run
our game. With the game speeding up as time passes and updating the game
state every so often.
Also started on the keyboard interupts. Have a circular buffer to hold
keyboard presses. It appears to work as desired but has not been heavily
tested because we need the handlers first.