Added supplementary page table struct to threads
Added a supplementary page table (spt) as a field of struct thread,
allowing us to keep track of the segments to be loaded into
memory lazily, i.e. only when a page fault occurs. The page
fault handler still needs to be modified to make use of the
additional data structure.