Implementing lookup in supplemental page table
Can now lookup based on a virtual address in the supplemental page
table. The function handles going to the next lowest page size so that
it can be used on any virtual address. Also began taking out ifdef VM
because we feel it is not necessary to keep reverse compatible with
this.