Stack growth from syscall does not work
Stack growth from the syscall should theoretically work. The problem is
in the page fault handler its getting a stack pointer at an address
higher than PHYS_BASE. Not sure what the issue is here. Could be that
you need to page fault before going into the file_read code but its
unclear. Also changed the STACK_FLOOR variable which is there to set a
lower bound on the stack. Right now its 100 times the PGSIZE from
PHYS_BASE, but this is pretty arbitrary, need to think about it more.