Fixed expanding stack from sycall
The probleem was that I couldn't led the code page fault in the syscall.
The solution is somewhat sketchy but it works. I made it such that if
according to the heuristic that the stack should be expanding, I just
have the syscall call a function from exception that will expnad the
stack from the address.